requirements.txt
python-telegram-bot
python-dotenv
emoji
forex-python
psycopg2
aiosqlite